+++ Dockerfile
@@ -1,4 +1,3 @@
-
 FROM python:3.6.3-jessie
 
 EXPOSE 80
