+++ example.env
@@ -1,6 +1,6 @@
 POSTGRES_DB=mlenabler
 POSTGRES_USER=mlenabler
-POSTGRES_PASSWORD=mlenabler
+POSTGRES_PASSWORD=
 POSTGRES_ENDPOINT=localhost
 POSTGRES_PORT=5432
 MLENABLER_TEST_DB=postgresql://admin@127.0.0.1/ml_enabler_test
